Add files via upload
In this task, I created a bar chart/histogram to visualize the distribution of a categorical and continuous variable. Specifically, the visualization represents the distribution of ages and genders within a specified population. The goal is to provide clear insights into the demographic structure, highlighting patterns and trends within the data.

Data Source: Kaggle
Tools Used: Matplotlib, Seaborn, Pandas
Visualization Type: Bar chart/histogram
This visualization aids in understanding the demographic distribution, making it easier to interpret the data at a glance.